Adding some small changes f

+++ security/src/main/java/com/ali/dao/repositories/QuestionRepository.java
@@ -1,8 +1,7 @@
 package com.ali.dao.repositories;
 
-import com.ali.dao.entities.Certificate;
 import com.ali.dao.entities.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Question, Integer> {
+public interface QuestionRepository extends JpaRepository<Question, Long> {
 }
